:truck: Reorganize as some library-ish file/dir structure
* bin, lib, test にファイルを分ける
* lib/bowling.rb でまとめて `require_relative` する
