+++ 07.bowling_object/bin/bowling
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+# frozen_string_literal: true
+
+require_relative '../lib/bowling'
+
+puts Bowling::Game.new(ARGV[0]).score
