+++ 07.bowling_object/lib/bowling.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require_relative 'bowling/score'
+require_relative 'bowling/rolls'
+require_relative 'bowling/parse'
+require_relative 'bowling/game'
+
+module Bowling
+end
